Attempt to create the main layout of Tracker.

+++ implementation/MangaCheck/data/tracking/src/main/java/it/unitn/disi/lpsmt/g03/tracking/TrackerSeries.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 
 @Entity("tracker_series", indices = [Index(value = ["title"], unique = true)])
 data class TrackerSeries(
-    @PrimaryKey(autoGenerate = true) val uid: Int = 0,
+    @PrimaryKey(autoGenerate = true) val uid: Long = 0,
     @ColumnInfo("title") val title: String,
     @ColumnInfo("status") val status: ReadingState,
     @ColumnInfo("description") val description: String?,
